+++ Source/DynamicalModels/ObjectModels/imstkRigidBodyModel.h
@@ -130,7 +130,7 @@ public:
             //auto q = Quatd(r);
             PxTransform pose;
             //pose.q = PxQuat(q.x(), q.y(), q.z(), q.w());
-            pose.p = PxVec3(p.x(), p.y(), p.z());
+            pose.p = PxVec3((float)p.x(), (float)p.y(), (float)p.z());
             m_pxDynamicActor->setGlobalPose(pose);
         }
