+++ Source/Geometry/Decal/imstkDecal.cpp
@@ -40,7 +40,7 @@ Decal::print() const
 double
 Decal::getVolume() const
 {
-    return m_dimensions.x * m_dimensions.y * m_dimensions.z;
+    return (double)m_dimensions.x * (double)m_dimensions.y * (double)m_dimensions.z;
 }
 
 void
